+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - **Breaking Change**: `Definition.refs` will be `Reference[]` instead of `string[]`.
 - **Breaking Change**: `RefIdGenerator` is `(ref: Reference) => string`.
 - **Breaking Change**: `RefRenderer` is `(ref: Reference) => string`.
-- **Breaking Change**: rename `BiParser.collectXXX` to `BiParser.parseXXX`.
+- **Breaking Change**: re-write `BiParser`.
+- **Breaking Change**: rename `BiDoc.collectDefinition` to `BiDoc.collectDefinitions`.
 - Fix: escaped reference for alias will be rendered as the alias instead of the name.
 
 ## v0.1.6